Break up elemwise_unary_op_basic.{cc,cu}
